changed how the follow list works
* show all follows instead of just live follows
* live streams are highlighted and at the top of the list
* separate, scrollable blocks for twitch and mixer

note: this new structure has a much higher potential for API calls (in order to retrieve the entire follow list + find live streams), and since the rate limit for the respective APIs is relatively low for a client that is not associated with a twitch or mixer account (read: somewhere below or around 100/min) the limit is easy to reach. this goes especially for twitch.